Add missing libraries.
Looks like the plugin is in general built without explicitly resolving
symbols. On my mac the compiler requires that libraries are fully bound
and therefore it complains. This is the set of libraries the checker seems
to require.

Whether this is the correct I'm not sure (I vaguely remember that
clang plugins need to resolve symbols when loaded) so the correct solution
might actually be forcing unbound libraries.